refactor: sboot web
