+++ jdevelops-framework/jdevelops-sboot-web/src/main/java/cn/jdevelops/sboot/web/context/ContextManager.java
@@ -1,0 +1,27 @@
+package cn.jdevelops.sboot.web.context;
+
+
+import cn.jdevelops.sboot.web.context.service.JdevelopsContext;
+
+/**
+ * 上下文全局
+ * @author tn
+ * @date 2023-02-11 22:50:15
+ */
+public class ContextManager {
+
+	/**
+	 * 上下文Context Bean
+	 */
+	private volatile static JdevelopsContext jdevelopsContext;
+
+
+	public static JdevelopsContext getJdevelopsContext() {
+		return jdevelopsContext;
+	}
+
+	public static void setJdevelopsContext(JdevelopsContext jdevelopsContext) {
+		ContextManager.jdevelopsContext = jdevelopsContext;
+	}
+
+}
