+++ jdevelops-framework/jdevelops-sboot-web/src/main/java/cn/jdevelops/sboot/web/service/package-info.java
@@ -1,0 +1,1 @@
+package cn.jdevelops.sboot.web.service;
