+++ jdevelops-framework/jdevelops-sboot-web/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,2 @@
+cn.jdevelops.sboot.web.scan.SwBeanInject
+cn.jdevelops.sboot.web.scan.SwBeanRegister
