--- jdevelops-webs/webs-springs/src/main/java/cn/jdevelops/springs/context/util/PathMatcherHolder.java
@@ -1,28 +1,0 @@
-package cn.jdevelops.springs.context.util;
-
-
-import org.springframework.util.AntPathMatcher;
-import org.springframework.util.PathMatcher;
-
-/**
- * 路径匹配
- * @author tn
- */
-public class PathMatcherHolder {
-    public static PathMatcher pathMatcher;
-
-    public PathMatcherHolder() {
-    }
-
-    public static PathMatcher getPathMatcher() {
-        if (pathMatcher == null) {
-            pathMatcher = new AntPathMatcher();
-        }
-
-        return pathMatcher;
-    }
-
-    public static void setPathMatcher(PathMatcher pathMatcher) {
-        PathMatcherHolder.pathMatcher = pathMatcher;
-    }
-}
