--- jdevelops-webs/webs-springs/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  cn.jdevelops.springs.scan.SwBeanRegister,\
-  cn.jdevelops.springs.scan.SwBeanInject
